Make test logger to reuse
